3 toboggan trajectory
Almost like in SkiFree, we travel and almost try to avoid trees in a
grid map but the toboggan gets some trees in the way so they're counted.
This would probably look embarrassingly easy with real nice matrix or
iter tools.

The second part must be with the map extending down too or something,
finding a better route. Or maybe it's too early for such difficulty.