4 passport processing
A valid passport is recognized by the fields that are present in it.
Parse multiline specs separated by two newlines. Figure out which
passports have all the necessary fields. Extra fields don't seem to be
harmful, and cid seems to be added just for some extra confusion, or
maybe it's an easter egg of some sort.