4b strict rules for automatic validation
Add the specified validity checks for each field and count the passports
whose fields pass the checks, not just the ones that have the required
fields with arbitrary data.

In a real world the passport records would be hashmaps instead of
vectors of the fields, but with the given input size, this will do.