6 custom customs
With just 26 questions to answer, the answer set of each person could
easily be a bitmap. With just a couple thousand people though, that
doesn't really matter so it's better to practice memorizing useful
standard library tools, like HashSet. Build an empty set per group and
extend it for each person in the group, and find the lengths of these
resulting sets.