7 handy haversacks
The input is still very small, but let's still build a proper graph to
search. Given a bunch of bags that contain other bags, build a network
of bag contents in reverse order to find which bag is inside other bags.
Because a bag color can have multiple possible hosts, this isn't a neat
DAG though, so keep track of the visited nodes. Otherwise just DFS.