8 handheld halting
Finally the traditional virtual machine exercise begins, and we start
with no conditional control flow and don't even run the programs to
completion. The sample is an infinite loop and the proper input probably
is as well. Advance the program counter and modify the accumulator until
a previously seen instruction is reached again.