9 encoding error
With a spec that looks like a hint about some interesting encoding or
encryption, unfortunately a simple brute force is enough with this input
size. Do as the mission says: for a window of 25 (or 5 for the example)
behind each value x, go through all pairs and see if there's a pair that
sums to x.

With some huge window, the number of identical additions would become
quite large and filling a storage of O(n^2) would help avoid duplicate
calculation. But that would be a lot of space for a large input size,
but the input isn't large. Could be if the window was huge though.