9b break xmas with the invalid number
The backstory is still a bit lacking in concreteness, but now we're
looking for a window of numbers that sums to the invalid number. This
time a brute force is too boring, so let's do something that in 2D would
be a summed-area table but is too easy in 1D and wouldn't even need that
extra storage, but is cool nevertheless. Scan the elements and produce a
new list of sums to each element; a window sum is then obtained by
subtracting two numbers.

Could just loop the two range ends and calculate the ongoing sum while
doing that, but this is a fun exercise for Iterator::scan.

This particular implementation style has a slight bug of not being able
to handle the front or back of the list, depending on indexing. Doesn't
matter here because the solution clearly is somewhere in the middle.
This is still AoC, not enterprise stuff.