10 adapter array
With the bag full of joltage adapters that have apparently been chosen
so that they can go all together, find the chain to join them all and
collect the 1-j and 3-j diffs. When the adapters can be considered a
network, is just arranging them from smallest to largest a topological
sort? Maybe the second part has something more interesting to do with
the graph, like the shortest path.

The first and last special cases - the wall and the device - of the
adapter chain are annoying corner cases and maybe the chain
representation should be the joltage deltas including them instead of
just the adapters. Oh well.