10b adapters in all the ways
Now the list of adapters really is a graph, and we're looking for the
number of paths in it from begin to end. DFS the graph and memoize the
results so that most of the paths that are very similar wouldn't get
recomputed. Looks like the list could actually be traversed in one pass
without the recursive fun, but oh well this is more generic.