11 seating system
"Fortunately, people are entirely predictable and always follow a simple
set of rules." lol what a world. Construct the specified cellular
automaton and run it until it stabilizes. I wonder how the maps are
generated so that the simulation will always stabilize. At least it
oscillates towards the stable state which is interesting.

Out-of-bounds lookup is always something I can't decide on. A hashmap
with keys of (x,y) would be handy, a vec of vecs is not bad with get()
and a flattened vec would be efficient but more tricky.