11b people care about visible seats
Now with the rules of the game changed slightly, rearrange the core
logic to support some generic rule function for computing the next cell.
The new way to compute seats considers seats also visible far away, not
just neighbors, and the people tolerancy level is also increased.

This is probably a function pointer now? A trait-level generic thing
might be easier to inline, but the runtime is still fast enough to not
care.

The result of 2047 is suspiciously close to 2^11.