12b the meaning of waypoint
Equipped with the instructions on how to read the instructions, we now
have a waypoint that resembles the mechanism of a spaceship from the
hitchhiker's guide etc but not the infinite drive thing but something
else that I can't put a name to, so that our ship warps towards a
waypoint and most of the actions process the waypoint. This would
probably make a fun animation.