13b shuttle company contest
Given the buses and their offsets, we get a system of congruences and
some time wasted on staring at lcm/gcd. This CRT solution had to be
looked up, there's probably some less fancy trick though but the sieve
was fun to implement in the morning based on just the math. No need to
order the numbers by decreasing moduli because this is fast already. Or
at least on my input, it is.

"The Chinese remainder theorem (expressed in terms of congruences) is
true over every principal ideal domain." mmhm thanks wikipedia