14b the v2 chip floats like a quantum computer
Now the X bits depict the tristate address lines left hi-z. Somehow they
get all the values instead of some random value, so loop through the
possible combinations of the bits, generating the numbers in between,
and for each number scatter the bits of the number to the X bits: for
each nth X bit take the corresponding ith bit in the source. Sadly just
a brute force translation is again good enough, so no clever bit
trickery is needed. On some processors there seems to be an instruction
for this though (PDEP).