15 rambunctious recitation
With game rules referring to previous numbers, keep track of the game
history and remember those numbers' turns. The difference between two
previously spoken turns or just the turn numbers themselves could
probably be stored directly, but store the two previous literally as
specified because it was early for my brain. Without the branching in
the loop this might be solvable analytically but those branches always
add corner cases in aoc that need the brute loop.