17 conway cubes
A cellular automaton in three dimensions is handy enough to store in a
hashset with the coordinate as the key. Nothing too fancy here, and the
dimensions are looked up again in non-constant time for every iteration.
Could grow them by one for every iteration because that's the max they
can expand, but this is less clever and thus more robust for this
morning's brain activity.