17b pocket dimension is 4d
Extend the game to work in four dimensions, and for the first part of
the puzzle, just discard everything but coordinates in the w=0
hyperplane at the end of each iteration so that the result stays valid.
Some extra computation is thus made for the border dimensions of w, but
that's the price of this lazy approach. Could extend the functions to
some trait thing around a generic coordinate. That should be a neat rust
exercise for another day.