18 operation order
With a math expression that has no specific operator precedence,
evaluation is easy enough by munching the tokens as they appear.
Subexpressions in parens are evaluated recursively and then processed
like numbers; the numbers and operators interleave naturally so it's
enough to store the result computed so far and the number currently
being parsed, if any.