18b advanced math: simple operator precedence
Now with two kinds of operators, precedence levels require just a slight
extension to calculate sums before multiplications: if a multiplication
(that has a lower precedence in this twisted world) is found,
recursively evaluate the rest of the expression first before evaluating
the multiplication. This crazy hack wouldn't scale but works beautifully
in this case.