19 monster messages
With rules that look like regexes, build a regex and let the regex crate
parse it instead of writing a proper parser. The rules are still parsed
to build the regex though... Luckily there indeed aren't any
self-referencing rules, or else the builder would go to an infinite loop
when trying to calculate what's the regex str representation of an
expanded rule.