19b matching rules aren't quite right
Hardcode a replacement for two rules as specified for two loops. Looking
carefully, these rules 8 and 42 are used only by the rule 0 which is
processed first, so this looping could be hardcoded in the main logic
too. Because we're cheating already with regexes, cheat some more by
unrolling the second loop at most 10 times. The first one enjoys native
regex support.

The sample input is modified slightly to work with the assumption that
indices don't overflow the number of rules. The sample is less than 43
long but refers to index 42 etc. The code doesn't hardcode the index 42,
only 8 and 11.