20 jurassic jigsaw
No bugs whatsoever for the first try, cool! Given the puzzle pieces,
parse only their names and borders. The borders have to match into some
solution that will be rotated or flipped in some manner but all such
alternatives are equally good for the corner pieces. Exhaustively DFS
all possibilities which there hopefully won't be too many: start with
one puzzle tile laid on the table, find what neighbors it can take, and
try each and recurse into their neigbors.

Looking at the animation of how this progresses, for a general solution
for larger maps it might make sense to subdivide the map / join multiple
smaller tiles, and flip or rotate or move collections of tiles that are
relatively large (perhaps rectangular, for simplicity). Now many
solution candidates are searched nearly halfway through the entire map
just to get completely discarded because the joined total would belong
to the other end of the map instead of where the search starts; finding
a nice half-square and keeping it could speed up the rest of the
process. Trying to match tiles at exact positions of the map instead of
more generally to each other is perhaps too restrictive and is certainly
not how humans assemble jigsaw puzzles; the square requirement could be
performed later in the search because specific border tiles don't exist.