20b check for sea monsters
Ok so that's what the jurassic means here. Once complete, the jigsaw
image represents a map of sea monsters (that apparently lie there on
their sides, or else this satellite is angled very unusually). Rotate or
flip the image such that the monsters are found; there's only one valid
orientation. Count the monster bit patterns, mask them away, and count
the number of ones in the remaining bits.

The tiles' images are rotated and flipped non-lazily while searching for
the puzzle solution. This unnecessary mangling is not currently a
performance problem but could be avoided by storing just the orientation
wrt. the original, and finally rotating each tile in just one go once
the final configuration has been found.