21b match ingredients and allergens
The input is again well-behaving and the initial mapping from allergen
to ingredients is small, and it contains at least one unique match to
begin with. With this match recorded and removed from all other
possibilities, another unique match (or more) appears. Loop this until
no remaining unknowns remain, and print the ingredients in the correct
order (by the allergen name).