21: extract the hashmap remove_by logic
This is a curious rust challenge so document it, and a separate function
(instead of those closures) makes the while let loop slightly neater as
well. The hashmap api appears to support removing an entry by key but
not by value. An entry can be found by value from the entry iterator api
but that returned k/v pair can't be used directly to remove the entry
because it's a ref to the data itself instead of some magical internal
element type of the hashmap, so remove_entry() has to look it up by the
key again.

Of course in this case any extra computation does not matter but the
code is a bit clumsy.