22 crab combat
A small crab learns to play cards. Pop cards from two decks (just vecs
with this small input) and collect them to the player who got the card
with better value. The game ends once one deck is empty. Looks like the
input is probably specifically chosen to be nice so the game terminates.