22b defend raft captain honor recursively
Incorporate the new rules of the game such that each game can spawn a
sub-game. Each game stores a history of all previous game states as
specified in the rules to look for cycles that result in the first
player winning. The recursion is simple enough; just the mission spec
was slightly confusing so this had initially a bug of not truncating the
decks for the sub-game (and this bug wouldn't affect the outcome of the
sample game!).