23 crab cups
The crabs here really like games. Today a game is about shuffling cups
around. The input is small enough so a vec works well, and vec even has
nice support for rotating the contents. It's handy to keep "the current
cup" always in the front so the rest of indexing is slightly easier.