23b one million cups
Not very fluent in crab. The number of cups is one million (our raft is
big?) and the game runs for ten million moves. Now the handy vector
rotations aren't feasible, so rewrite the game with a linked list
instead. The list needs to be just singly linked and since the game
consists of a nice list of consecutive integers, the list nodes can be
indexed by node in a flat plain old vector.

Curiously this was much (20x) slower with a hashmap initially; maybe its
storage is less efficient for this purpose, although the access pattern
seems so random that a linear vector shouldn't help that much.

Spent quite some time looking for a cycle here. Long loops always seem
like they'd require looking for cycles and matching to it, but ten
million isn't actually a big number. A trillion would be.