2021/1 sonar sweep
Produce consecutive pairs of the input values and find which dive deeper
in the sea floor.

The iterator and stuff library didn't quite happen last year as planned,
but this year one nice practice could be to overengineer all these to be
more generic than enough, just to learn rust generics/traits. Some
proper error handling would also be useful practice but it would be too
much work to test reliably because the input data is always good.