4 giant squid
Parse the the bingo configuration and play until one player gets one
full row, because what vessel does not have a bingo subsystem??
Curiously a squid _outside_ the vessel is considered a passenger, maybe
it'll pay for the ticket. A board is defined by an array of numbers and
corresponding flags on which numbers have been marked. The array is a
flattened representation of five elements in each direction.

The logic assumes that the game input is well-formed; only one player
should win at a time, and some player must win before the numbers run
out.

Parsing this from the input is also surprisingly easy with just chunked
sections, with the board lines flat_map'd to one series of numbers.