unify 2a and 2b
Turns out that the turning of the ship still keeps enough information
for the score of the first puzzle, so delete the duplicate code that
calculated the simple score. Run the sub only once, and figure out both
scores at the end; the x coordinate is identical in both, and the y
coordinate of part 1 is equivalent to the aim from the corrected
instructions.

A proper level of overengineering for this day would be to parse the
instructions to some well typed enums, but come on there's just three.
Or perhaps the instructions could define a microarchitecture for the sub
that would be JITed?