cleanup 3, more generic and partitioning
The bit criteria that flips the ones-or-zeros selection for the bit
criteria can of course be used in the same condition like so.

Also, while this partitioning does require mutable access to the input,
it makes the logic cleaner and perhaps faster as the data size reduces
to about half on each iteration. The order of the diagnostic report has
no effect to the results.

This makes me still wonder if both the ratings could be computed in a
single loop with even more clever partitioning.