generalize bingo for any board size
Just to practice the new const generics, make the bingo board generic
over the board size. The squid game logic with its unusual rules is
still enforced to use the size five.

As instructed by the compiler, an empty-looking bound is required on the
struct and then the impl: "where [(); N * N]:". Thanks compiler, have to
go read about that "unconstrained generic constant".