5 hydrothermal venture
Given a list of vent ranges in lines, parse the line coordinates and
loop the integer coordinates that each span. For each visited
coordinate, mark in a map that one more vent has been found.

For those loops to work properly, sort the line coordinates so that
they're increasing.

Parsing like this is terrible but simple; I wouldn't like to use another
crate for such a simple input.