5b also diagonal lines
This time loop all lines, walk along the line length from start to end
and note both x and y coordinates. THe 1:1 diagonality sure helps.

Lots of existing things have to change because sorting the x and y was
not correct, and i32 is more convenient for the coords because their
diff can go negative which is nice for the new loop.