6 lanternfish
As the spec hints about exponential growth, avoid keeping an entry for
each fish because their count is going to explode. Because the fish
lifetimes are integers and fairly limited in terms of growth logic,
store the amount of fish for each age between 0 and 8 inclusive. During
each iteration, move the ages - or the time until reproduction - back,
and sum the reproduced count to the init age.

The copying loop looks like copy_from_slice but something even more
clever looks possible; this just was easy enough in the morning.