6b would they take over the ocean
Insight on the exponential growth was helpful; just do as instructed and
see what the population is after 256 iterations because the calculation
can handle that; 256 isn't that much more than 80. With the sample it
would be gigabytes of fish. The fish probably would not take over the
entire ocean in real life because they'd run out of food.