9b the largest basins
Do some embarrassing copypasta for finding the lowest points, and then
DFS each low point for its non-nine neighborhood to find the matching
basin size.  This is possible because the basins are nicely isolated
from each other with heights of 9, which is also easily seen from the
input with a little highlighting in a text editor. With the basin sizes
computed, sort them and get the top three.