11 dumbo octopus
Cute creatures are arranged nicely in a grid. Iterate them in steps and
simulate how they would flash, and how the flashes propagate recursively
in the octo network. Track the flash count during the simulation. Lots
of off-by-one candidates in this logic.

Because the sample is provided with a nice visualization, implement
identical visualization to be able to compare the output in case of any
bugs, which there were initially with the "<= 9" condition.

Indexing these vecs is again nasty with the type conversions but the
conversions are acceptable now with the signed neighbor coordinates.
(Maybe the i32 should be isize though to be pedantic.)