12 passage pathing
The sub is not subbing well so find a path using a map. The input is a
list of undirected graph edges, so build graph nodes and edges in both
directions. DFS the graph with a state at each search step of which
(small) caves have been visited, as they must only be visited once. Keep
track of the global number of paths visited (which could also be a
return value of the search, but this happened first).

The input appears to be well formed such that no infinite loops exist
between two big caves, so no extra logic needs to be done for such
conditions.