generalize cave search logic to avoid duplication
The non-flexible search where there's no time for an extra visit is
basically the same as the flexible one as if the double visit would have
been done already. The previously visited node is not actually needed to
keep, so keep just a boolean flag of whether there's time for one extra
visit of a small cave. The two searches differ by just the starting
value of this flag.