13 transparent origami
Origami usually makes 3D structures but now we're just folding a paper
flat. Parse the input and fold it once, removing the points that would
lie on the folding part and adding them back to the mirrored coordinates
along the fold axis. Keep the points in a set to not worry about
duplicates.