14 extended polymerization
Help reinforce the submarine with a polymer template and pair insertion
rules: bruteforce the ten first steps and count the elements. The length
looks to be obviously growing exponentially so this certainly won't do
for the next problem, but at least it's a good trivial reference
implementation to compare against.