14b stronger polymer
That 2^40 is too much to work with hardcoded polymers, so encode just
the counts of the pairs; their order is not important. The puzzle answer
on element counts needs some care because the pairs include each element
twice (or more appropriately half of each element twice, if you think it
that way) except the first and last in the chain. (Thanks Peter K.)