15b cave is five^2 times larger
The original scanned area repeats in a consistent way, so build the real
map from the scan by inserting the five by five coordinates from each
scanned coordinate and then use the same search algorithm with this map.
The decision to use a hashmap now really pays off because indexing a vec
of vecs would again be rather unpleasant even though the map is dense
and hashing dense coordinates is a bit silly.

The four level flat_map is not the most readable when compared to plain
nested for loops, but at least it's fun practice. Collecting a hashmap
from an iterator is very convenient.