16 packet decoder
A transmission from the elves is sent using BITS. Ha. The packet data
does not align with the four-bit elements in any way, so keep track of
bit and element indices and extract just the right amount each time.
Walk the specified packet format recursively and collect the version
sums. When recursing, make a substream of the original stream by copying
just the offsets and by specifying a length; this way the bit data does
not need copying.

A more generalized packet walk might be more suitable for reuse, but
it's early in the morning.