16b calculate the transmission expression
Copy the recursion but now keep computing the value of the
subexpressions. Fill in the operator kinds and execute them with the
values collected from the subpackets.

The subpacket values are collected in vectors which might be wasteful in
allocations when an iterative computation might also work, but this is
more ergonomic for now and perf does not really matter.

This is complex enough to look like a reusable piece of a machine to be
used in a later day like the VM stuff from previous years; fingers
crossed, as that would be fun.