17 trick shot
Simulate the probe trajectory by iterating it based on the very accurate
physical model. Run the simulation for quite high numbers of starting
velocity until the probe falls inside the target area's y coordinates,
with some insight on the number limits gathered manually.

Note that the starting y velocity appears to be nearly the same as
smaller y of the target. This hints that the probe is moving at the same
rate but just different direction when it's traveling back down. The y
velocity is not even affected by wind. Brute force the limit anyway.