18 snailfish
A school of fish obviously has math homework. Snailfish numbers are
recursive pairs, so parse them into a tree structure with way too many
boxes; this is over a million allocs according to valgrind.

Parse the numbers recursively and run the sum of all numbers. Splitting
is an easy task as it does not affect other elements but just one
element is replaced; the explosion is more involved. Carry the explosion
"fragments" upwards the call stack in return values until they can be
used (i.e., the caller has elements in the correct direction), if at
all. If the caller has those, recurse again into the nearest element in
that direction and apply the addition.

With a much larger set of computation some arena might be nice for the
individual regular numbers. For now this is fine as a pure math problem.