18b a second question on the back
Always flip the exam paper over. Doing the calculation with the parsed
data structure is not so bad, so just brute force sum all the pairs of
different input numbers and find the maximum.

There probably exists some trick to reduce the input values first into
some meaningful smaller pieces of data so that the magnitude would still
be valid, or the tree could be flattened to a nicely ordered array
because it has a maximum depth. Or ignore the tree and (ab)use a list?

Lol valgrind runs 45 seconds and reports 20 million allocs. 0.6 s for a
native run is a bit on the slow side too.