19 beacon scanner
The scanners are excellent beacon locators but do not locate themselves.
Take the first scanner as the one true coordinate system and compare
each of the remaining scanners to it, rotated in every possible
orientation. When the orientation happens to match that of the master
coordinate system, the points look identical but the set is just shifted
somewhere based on where the new scanner is. When the scanners see a
large enough shared subset of points, any of those shared points can be
chosen as a common origin to which each other point can be referenced.
The subset is not known in advance, so try every possible pair of origin
beacons in the master set and the candidate neighbor set. When a match
is found, transform the new points back to the master scanner.

For now this is terribly inefficient because the results are not cached
in any way and the same calculation needs to repeat over and over.