20 trench map
Mapping the ocean trench involves a proper CSI algorithm that enhances
an image with a convolution kernel so that it gets bigger and bigger on
the edges. Increase the image size on each iteration. The trick is to
keep track of the infinite background type that affects the edges
because with this input, the 3x3 kernel from light pixels goes to dark
and vice versa; it blinks.