21 dirac dice
Implement the game logic in a fairly hardcoded manner to be able to
understand it and to prepare for a possibly different dice, because the
practice game is explicitly said to use deterministic dice. Just loop
the game logic until a score is more than 1000.

The circular game track is again more handy to work on in modular
arithmetic, so be careful to use zero-based internal indexing and then
translate to one-based when calculating the scores.