21b play the real game
Preparing for changes by keeping the game loop simple was not effective,
as the game is now entirely different as each die roll triggers new
universes and the puzzle calls for the win counts only.

Keep the win counts memoized in a map of states: from the begin state
with the players in some positions, new game states emerge until the
game ends at a score of 21 on either player, which is achieved in just a
few rounds. This forms a recursion that ends at the game end, and
produces fibonacci-like repeating subsearches whose results can be
memoized. The entire state is only 10 * 10 * 21 * 21 position and score
combinations or so, and that loops through fairly quickly.

The computation for each player is done separately which is a bit silly;
indexing by player id would make this much more succinct.

I spent ("wasted") some good amount of time analyzing how the universes
could actually be enumerated such that each "universe ID" would encode
the entire history of the game in base-3 trits of die rolls, so each
universe knows what the game state is. The game history is not important
for the result though, but that felt interesting to look at and it
looked vaguely similar to z-order curves.