22b time to reboot the reactor
Now with the humongous cuboids the bruteforce loop is no longer an
option, so do the calculation based on counting entire cuboids' volumes
that are easy to compute. Iterate all the cuboids and add their effect
on the space one by one and keep cuboids in the space non-overlapping
and positive ("on").

When one cuboid is added, any intersecting space with another existing
cube needs to be removed, and the resulting irregular shape will consist
of smaller cubes for which the insertion is executed recursively again
until the new cuboid does not intersect with anything and it's added
as-is.

When one cuboid is removed, its volume will get subtracted from each
cuboid that overlaps it; this way the existing volumes will get split
into irregular shapes as above. They need no recursion though; the
original box never overlapped another so the new ones won't either.

The CSG subtraction is done in the most straightforward manual way
possible. The cutting cuboid may cover the entire volume, be completely
inside it, or touch any number of faces. Assume that the cutting box is
somewhere inside and build all 3^3 volumes around it based on some
visual thinking, then discard the empty volumes that may happen when the
cutter was not inside as seen from one face. A much better way to do
this CSGing could be to split the volume in one axis at a time on a face
of the intersection box and to recurse the results; that would probably
mean less calculation to do and the amount of code would reduce
superlinearly.