23 amphipod
A group of amphipods share a burrow that happens to be very regularly
shaped and they need to move into designated corridors. Find a way.

The burrow forms a very simple graph, so just hardcode the graph
structure as a world map instead of spending time debugging code to
parse the input into a general graph. The search of the optimal homing
jostle forms another implicit graph where the edges are defined as the
stubbornity rules and the possible movement of the amphipods as
determined by the world map. The states in this graph can be coded as
the locations of the amphipods, although an entire occupancy map of the
world map could work too as the map is relatively dense.

With that implicitly defined graph, run a dijkstra but with a heuristic
so it is actually a* search. The goal state does not quite work like
that because of the ambiguity of who's who in a group of pod friends but
the search finds a good spot anyway based on the heuristic of remaining
moving cost going to zero. Probably something wrong because the search
does not always find the optimal number.