24 arithmetic logic unit
This ALU is a bit special, it can also do input and stuff. Simulate the
model number checker program in a virtual machine by executing the
instructions as specified (fortunately, the program never divides by
zero). Brute-forcing fourteen-digit numbers is not feasible even if the
program would be converted to native code, so reverse engineer the
program logic for good numbers and run the program just for
verification. The program has rigid structure so pattern-match the
structure and print out the important numbers for some manual finish.

The program uses the z variable as a stack of base-26 numbers; it pushes
some of the the input digits with an offset value and pops off some in
certain conditions. With a MONAD-conforming model number, it pops
exactly all pushed values out, leaving zero in the z variable and thus
nothing in the stack. In a valid MONAD, each popped value needs to have
a specific relation with the most recently pushed value. Specifically,
some digits need to be equivalent to other digits plus a per-digit
offset. (Looking at the z value at the end and backtracking by hand
quickly turned out to be not feasible.)

The maximum valid model number has nines in all digits where a nine
fits, and nine minus an offset for the digits that depend on another
digit in a restricting manner. A nine fits in a digit where its pair
would not become larger than nine (e.g., when digit 1 = digit 0 + 1, the
max for digit 0 would be 8.).

(There are 26 characters in a-z, coincidence or an easter egg or
deliberately confusing spec? Are the values in the stack supposed to be
letters?)