23b the diagram was folded up
Discover an extra part of the diagram (another input file for now). Fix
a few bugs in the search algorithm and perform a few optimizations so it
runs in reasonable time, and expand the understanding on the new world
with deeper homes that fit four amphipods.

Adjust the cost heuristic for more exact data with the new graph.

Provide another hand-drawn map for the bigger world (FIXME: param it and
skip the big homes if they aren't needed). Print states with all
positions.

Save some memory in the state: make x and y only i8, although this is
not necessarily the best way to store the state wrt. optimization; the
graph is so dense that it could make sense to just store each node's
occupants in order. But with the below stuff the size is not even a
problem anymore. Oh well.

Trace the state history with scores to debug and verify the moves.

Fix a bug that resulted from stupidly reading that "from hallway to
room" rule in the spec to the letter; a direct room to room move can be
possible sometimes, but that does technically visit a hallway first;
such moves did not take into account the timid rule of moving only to a
safe room. So as a fix, don't go from room to room directly in one go.
Stopping in the hallway is equally expensive so it's also valid. (Could
also fix the "from hallway to room" check to be just "to room".)

The search wouldn't always find the optimal route, but one that was just
a couple of steps off. Fix this bug by not always stopping the neighbour
map DFS on a node that was visited; visit again if the new cost is
lower. That made some pods take detours when moving up from a room, as
those intersections have three nodes visible from each other. That DFS
was meant to be BFS but something happened in between.

Use an extra pruning rule for staying put when in a final position: when
in the correct home pipe and stacked on top of friends or at the bottom,
don't go anywhere. Moving from that state can only increase the cost
with no benefit.

Also, drop to a room only into a spot that is the final one, at the end
of a safe room that has been tightly filled. This too helps limit the
search space and works in combination with the above rule.

This patch does way too many things at once, but they all were written
in one go until the second part ran fast enough so not splitting this
keeps the work faithful to the original.