cleanup 23 from most of that hackiness
Make the search actually stop when a goal configuration has been found.
The nature of this dijkstra/a* makes the search find the optimal path
first, so exit when the combined heuristic cost is zero, i.e., when all
amphipods have found their homes. The goal state is not unique due to
many amphipods sharing the same type, so giving the search only one goal
state with exact locations for every amphipod is not feasible.

Strip off scattered debug stuff. Generalize easy bits out to avoid
duplication, and make the single input file work for both parts of the
puzzle. This is still a bit of a mess from the hasty iterative
development.