5 supply stacks
Parse a looks-tricky-but-isnt spec of stacks of crates and a list of
instructions to manipulate the stacks. Manipulation then is much
simpler, just run each instruction to pop from one stack and push to
another. The result is the top of each stack concatenated (nonempty
stacks guaranteed). This shuffling of items looks much like the towers
of hanoi game.