5b the cratemover is 9001
With air conditioning now in place we find out that multiple crates move
at once, so for each move instruction the moved crates do not reverse
but the moving stack stays as-is. Copy the top elements instead of
one-by-one pop.

To use destvec.extend(&sourcevec[pos..]) gets a bit involved to satisfy
the borrow checker with split_at_mut(). The first part of the puzzle
could actually use this too, and just reverse the source vec iterator.